Refactor file structure to work with package manager
