+++ oligos_prediction/main.py
@@ -1,0 +1,13 @@
+from .predictions import run_predictions
+
+if __name__ == "__main__":
+    # Example input sequence and synthesis scale
+    # sequence = "ATGCATGCATGC"  # Replace with your input sequence
+    # sequence = "ACUUTATUCCAAAGGGCAGCUGA"  # Replace with your input sequence
+    sequence = "CGAAGCGCCCTACTCCACTCCUGGACAUUCAGAACAAGAA"  # Replace with your input sequence
+    synthesis_scale = 0.5  # Replace with the synthesis scale (if applicable)
+
+    sequence_data = [{'sequence': sequence,
+                    'synthesis_scale': synthesis_scale}]
+    # Run the main function
+    predicted_yield = run_predictions(sequence_data)
