+++ pyproject.toml
@@ -2,5 +2,9 @@
 name = "oligo-predictor"
 version = "0.1.0"
 
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
 [tool.setuptools]
 py-modules = ["oligo_predictor"]
